1.18.2

+++ net/minecraft/world/entity/monster/Zombie.java
@@ -157,7 +157,7 @@ public class Zombie extends Monster {
 
    protected int getExperienceReward(Player p_34322_) {
       if (this.isBaby()) {
-         this.xpReward = (int)((float)this.xpReward * 2.5F);
+         this.xpReward = (int)((double)this.xpReward * 2.5D);
       }
 
       return super.getExperienceReward(p_34322_);
